adds stake_sol instruction

+++ programs/stacking_program/src/state/user_account.rs
@@ -4,8 +4,8 @@ use anchor_lang::prelude::*;
 #[derive(InitSpace)]
 pub struct UserAccount {
     pub points: u64,
-    pub nft_staked_amount: u32,
-    pub spl_staked_amount: u32,
-    pub sol_staked_amount: u32,
+    pub nft_staked_amount: u64,
+    pub spl_staked_amount: u64,
+    pub sol_staked_amount: u64,
     pub bump: u8,
 }
